requirements.txt
typer[all]==0.9.0
PyGithub==1.59.1
openai==1.3.7
jinja2==3.1.2
weasyprint==60.2
requests==2.31.0
rich==13.7.0
python-dotenv==1.0.0
pydantic==2.5.0
